make dataset a DB entity and add management commands, DB-based logger.

+++ spendb/model/common.py
@@ -1,4 +1,22 @@
+from json import dumps, loads
+from sqlalchemy.types import Text, MutableType, TypeDecorator
+
 from spendb.core import db 
+
+class JSONType(MutableType, TypeDecorator):
+    impl = Text
+
+    def __init__(self):
+        super(JSONType, self).__init__()
+
+    def process_bind_param(self, value, dialect):
+        return dumps(value)
+
+    def process_result_value(self, value, dialiect):
+        return loads(value)
+
+    def copy_value(self, value):
+        return loads(dumps(value))
 
 class TableHandler(object):
 
@@ -28,7 +46,8 @@ class TableHandler(object):
         bind.execute(q)
 
     def _drop(self, bind):
-        self.table.drop()
+        if bind.has_table(self.table.name):
+            self.table.drop()
         del self.table
